move grpc healthcheck to a separate package
